requirements.txt

# --- Core Deep Learning ---
torch>=2.0.0
# --- Core Deep Learning ---
torch>=2.0.0
accelerate>=0.27.0
transformers>=4.37.0
peft>=0.11.1
trl>=0.8.0
bitsandbytes>=0.43.0
sentencepiece>=0.2.0
safetensors>=0.4.0

# --- Data Handling ---
datasets>=2.20.0
pandas>=2.1.0
numpy>=1.26.0
tqdm>=4.66.0
requests>=2.31.0
beautifulsoup4>=4.12.2

# --- Evaluation Metrics ---
evaluate>=0.4.2
scikit-learn>=1.4.0
rouge-score>=0.1.2
nltk>=3.9.1

# --- Visualization / Notebook ---
matplotlib>=3.8.0
jupyterlab>=4.1.0
notebook>=7.0.0

# --- Utility / Packaging ---
huggingface-hub>=0.22.0
python-dotenv>=1.0.0
